visualiza-platform/backend/pyproject.toml
[project]
name = "visualiza-backend"
version = "1.0.0"
description = "Visualiza Platform Backend — AI-powered renovation visualizer for DR real estate"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "aiomysql>=0.2.0",
    "cryptography>=42.0.0",
    "redis>=5.0.0",
    "celery[redis]>=5.4.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "stripe>=10.0.0",
    "boto3>=1.35.0",
    "Pillow>=10.4.0",
    "httpx>=0.27.0",
    "scrapling>=0.2.9",
    "python-slugify>=8.0.0",
    "reportlab>=4.0.0",
    "jinja2>=3.1.0",
    "emails>=0.6.0",
    "tenacity>=9.0.0",
    "structlog>=24.0.0",
    "greenlet>=3.0.0",
    "python-dotenv>=1.0.0",
    "pymysql>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "factory-boy>=3.3.0",
    "faker>=28.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "python-dotenv>=1.0.0",
    "pymysql>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ANN"]
ignore = ["ANN101", "ANN102"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=80"
